UnmarkCommand: add support for unmarking tasks
Previously, users could only mark tasks as done, but there was no way to unmark a task as not done. This limited the flexibility of the task management system, as users could not correct mistakes or update the status of tasks that were marked as done by accident.

This commit introduces the UnmarkCommand, which allows users to unmark a task as not done. The following changes were made:
* Added UnmarkCommand class to handle the unmark operation.
* Updated Parser to recognize the unmark command and return an UnmarkCommand instance.
* Added markAsNotDone method in TaskList to support unmarking tasks.
* Updated Task to include markAsNotDone and added assertions for state transitions.
* Enhanced Ui to provide feedback when a task is unmarked.
* Added assertions in relevant classes to document important assumptions.
* Added/updated tests to cover the unmark functionality.

With this change, users can now unmark tasks, improving the usability and robustness of the application.